// src/app/api/models/usuario-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UsuarioResponse {
  activo?: boolean;
  correo?: string;
  id?: number;
  institutionId?: number;
  nombreCompleto?: string;
  nombreUsuario?: string;
  rol?: 'MASTER' | 'ADMIN' | 'CAJERO' | 'CLIENTE' | 'SISTEMA' | 'TUTOR' | 'CONSULTOR';
}